requirements.txt
# Gateway
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
PyNaCl==1.5.0
requests==2.31.0
PyJWT==2.8.0
email-validator==2.1.0
prometheus-client==0.19.0
python-multipart==0.0.6
boto3==1.34.0

# SDK
# (requests and PyNaCl already listed above)

# Policy Compiler
PyYAML==6.0.1
Jinja2==3.1.3

# Demo
rich==13.7.0
stripe==7.11.0

# Testing & Development
pytest==8.0.0
pytest-asyncio==0.23.3
httpx==0.26.0
locust==2.20.0
